V1.5.1

+++ database/migrations/2020_02_25_042103_fkads.php
@@ -13,16 +13,17 @@ class Fkads extends Migration
      */
     public function up()
     {
-	    Schema::create('ads', function(Blueprint $table){
-		$table->string('fk_name', 30);
-		$table->foreign('fk_name')->references('name')->on('users')->onDelete('cascade');
+      if (!Schema::hastable('ads'))
+  	    Schema::create('ads', function(Blueprint $table){
+      		$table->string('fk_name', 30);
+      		$table->foreign('fk_name')->references('name')->on('users')->onDelete('cascade');
 
-	    	$table->string('uri', 255);
-		$table->string('url', 255);
-  
-		$table->date('updated_at');
-	    	$table->date('created_at');
-	    });
+  	    	$table->string('uri', 255);
+      		$table->string('url', 255);
+
+      		$table->date('updated_at');
+  	    	$table->date('created_at');
+  	    });
     }
 
     /**
